fix(login): force dynamic SSR to fix CSP nonce blocking hydration
Extract LoginClient.tsx as the client component and replace page.tsx
with a thin server wrapper that exports `dynamic = 'force-dynamic'`.

Without this, Next.js statically generates login.html at build time.
The pre-built HTML contains inline React Flight scripts with no nonce
attribute. When middleware injects a strict nonce-only CSP at response
time, the browser blocks those unnonce'd scripts, React hydration never
runs, and Lighthouse sees NO_FCP (no content painted in 31 seconds).

With force-dynamic, the page is server-rendered per request so
app-render.js can extract the nonce from the incoming CSP request
header (set by middleware via NextResponse.next({ request: { headers } }))
and stamp it on every generated inline <script> before sending HTML.

Resolves the Performance Audit Lighthouse NO_FCP failure in CI.
